duui-sentencizer-corenlp/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==21.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
corenlp-protobuf==3.8.0
deprecation==2.1.0
dkpro-cassis==0.8.0
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
idna==3.4
importlib-resources==5.4.0
lxml==4.9.3
more-itertools==8.12.0
packaging==23.2
protobuf==3.20.3
pydantic==2.4.2
pydantic-core==2.10.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stanford-corenlp==3.9.2
starlette==0.27.0
toposort==1.7
typing-extensions==4.8.0
urllib3==2.0.6
uvicorn==0.23.2
zipp==3.17.0
